# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Tài liệu API",
        "contact": {
            "email": "minhminhb2vtt@gmail.com.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/bills": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lấy danh sách đơn hàng của người dùng",
                "operationId": "33456436f9b48d4aee609fa37fb0611a",
                "responses": {
                    "200": {
                        "description": "Danh sách đơn hàng của người dùng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bills": {
            "post": {
                "tags": [
                    "Bills"
                ],
                "summary": "Tạo hóa đơn",
                "operationId": "aa6d8159ba951369b80cf4d40d5254f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "size",
                                    "shoe_id"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "size": {
                                        "type": "interger",
                                        "example": 40
                                    },
                                    "shoe_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bills/{id}": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Lấy thông tin hóa đơn",
                "operationId": "4f2af7e00bd4ea08b689fda375bab2d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin hóa đơn"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bills"
                ],
                "summary": "Xóa hóa đơn chưa thanh toán",
                "operationId": "1d3fab5897de74e91fb1063013e00f65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Xóa thành công"
                    }
                }
            }
        },
        "/api/bills/{id}/pay": {
            "post": {
                "tags": [
                    "Bills"
                ],
                "summary": "Thanh toán hóa đơn",
                "operationId": "b3a7052aa57d8233e737869615764eda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thanh toán thành công"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy danh sách danh mục",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Danh sách danh mục"
                    }
                }
            }
        },
        "/api/shoes/{id}/comments": {
            "get": {
                "tags": [
                    "Shoes"
                ],
                "summary": "Lấy danh sách bình luận của sản phẩm",
                "operationId": "b392300a5da733f2df175589fa1c779f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách các bình luận"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shoes"
                ],
                "summary": "Tạo bình luận cho sản phẩm",
                "operationId": "56cd5cb24e6b273f54bf6cbc5545d73b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Good"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo bình luận thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Xóa bình luận",
                "operationId": "b683c30a239aa780ffc7e9b219e90811",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã bình luận",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Xóa bình luận thành công"
                    }
                }
            }
        },
        "/api/ratings": {
            "post": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Tạo đánh giá",
                "operationId": "5cd8244a0d463d65c3d4aff9bb590d0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "star",
                                    "shoe_id"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Good"
                                    },
                                    "star": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "shoe_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo đánh giá thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/shoes": {
            "get": {
                "tags": [
                    "Shoes"
                ],
                "summary": "Lấy danh sách sản phẩm",
                "operationId": "468cd78c6c3e410147b46e919026e0f8",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tên sản phẩm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Giá sản phẩm tối thiểu",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Giá sản phẩm tối đa",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Danh mục của sản phẩm",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm"
                    }
                }
            }
        },
        "/api/shoes/{id}": {
            "get": {
                "tags": [
                    "Shoes"
                ],
                "summary": "Lấy thông tin của 1 sản phẩm",
                "operationId": "c35798328800e64c43436059869efddf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mã sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin sản phẩm"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Đăng ký tài khoản",
                "operationId": "482fd8cbb98dc3225fa0e7ac40c1abe6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "name",
                                    "email",
                                    "address",
                                    "phone"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "nguyenvana"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nguyenvana@gmail.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "TPHCM"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Đăng nhập tài khoản",
                "operationId": "afd167661891a2c4c1b49a04643b9223",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công"
                    },
                    "401": {
                        "description": "Đăng nhập thất bại"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lấy thông tin tài khoản",
                "operationId": "85d739158552477973430d9c6aaf191e",
                "responses": {
                    "200": {
                        "description": "Thông tin tài khoản"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Đăng xuất tài khoản",
                "operationId": "9872a1ded930c0faa99c7f184b4cd654",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Bills",
            "description": "Bills"
        },
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Shoes",
            "description": "Shoes"
        },
        {
            "name": "Comments",
            "description": "Comments"
        },
        {
            "name": "Ratings",
            "description": "Ratings"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}